+++ code/lib/ful/intrinsics_x86.hpp
@@ -21,16 +21,16 @@ namespace ful
 		ful_target("bmi") ful_inline unsigned int andn(unsigned int x, unsigned int y) { return __andn_u32(x, y); }
 
 		// not and
-		ful_target("bmi") ful_inline int andn(int x, int y) { return andn(static_cast<unsigned int>(x), static_cast<unsigned int>(y)); }
+		ful_target("bmi") ful_inline int andn(int x, int y) { return static_cast<int>(andn(static_cast<unsigned int>(x), static_cast<unsigned int>(y))); }
 
 		// bit scan forward
-		ful_inline unsigned int bsf(int x) { ful_expect(x != 0); return __bsfd(x); }
+		ful_inline unsigned int bsf(int x) { ful_expect(x != 0); return static_cast<unsigned int>(__bsfd(x)); }
 
 		// bit scan forward
 		ful_inline unsigned int bsf(unsigned int x) { return bsf(static_cast<int>(x)); }
 
 		// bit scan reverse
-		ful_inline unsigned int bsr(int x) { ful_expect(x != 0); return __bsrd(x); }
+		ful_inline unsigned int bsr(int x) { ful_expect(x != 0); return static_cast<unsigned int>(__bsrd(x)); }
 
 		// bit scan reverse
 		ful_inline unsigned int bsr(unsigned int x) { return bsr(static_cast<int>(x)); }
@@ -39,13 +39,13 @@ namespace ful
 		ful_inline unsigned short bswap(unsigned short x) { return __rolw(x, 8); }
 
 		// byte swap
-		ful_inline short bswap(short x) { return bswap(static_cast<unsigned short>(x)); }
+		ful_inline short bswap(short x) { return static_cast<short>(bswap(static_cast<unsigned short>(x))); }
 
 		// byte swap
 		ful_inline int bswap(int x) { return __bswapd(x); }
 
 		// byte swap
-		ful_inline unsigned int bswap(unsigned int x) { return bswap(static_cast<int>(x)); }
+		ful_inline unsigned int bswap(unsigned int x) { return static_cast<unsigned int>(bswap(static_cast<int>(x))); }
 
 		// move data from string to string
 		ful_inline void repmovf(const unsigned int * src, unsigned int cnt, unsigned int * dst)
@@ -65,41 +65,41 @@ namespace ful
 		ful_inline unsigned int rotl(unsigned int x, int shift) { return __rold(x, shift); }
 
 		// rotate left
-		ful_inline int rotl(int x, int shift) { return rotl(static_cast<unsigned int>(x), shift); }
+		ful_inline int rotl(int x, int shift) { return static_cast<int>(rotl(static_cast<unsigned int>(x), shift)); }
 
 		// rotate left
 		ful_inline unsigned long rotl(unsigned long x, int shift) { return _lrotl(x, shift); }
 
 		// rotate left
-		ful_inline long rotl(long x, int shift) { return rotl(static_cast<unsigned long>(x), shift); }
+		ful_inline long rotl(long x, int shift) { return static_cast<long>(rotl(static_cast<unsigned long>(x), shift)); }
 
 		// rotate right
 		ful_inline unsigned int rotr(unsigned int x, int shift) { return __rord(x, shift); }
 
 		// rotate right
-		ful_inline int rotr(int x, int shift) { return rotr(static_cast<unsigned int>(x), shift); }
+		ful_inline int rotr(int x, int shift) { return static_cast<int>(rotr(static_cast<unsigned int>(x), shift)); }
 
 		// rotate right
 		ful_inline unsigned long rotr(unsigned long x, int shift) { return _lrotr(x, shift); }
 
 		// rotate right
-		ful_inline long rotr(long x, int shift) { return rotr(static_cast<unsigned long>(x), shift); }
+		ful_inline long rotr(long x, int shift) { return static_cast<long>(rotr(static_cast<unsigned long>(x), shift)); }
 
 # if defined(__x86_64__)
 		// not and
 		ful_target("bmi") ful_inline unsigned long long andn(unsigned long long x, unsigned long long y) { return __andn_u64(x, y); }
 
 		// not and
-		ful_target("bmi") ful_inline long long andn(long long x, long long y) { return andn(static_cast<unsigned long long>(x), static_cast<unsigned long long>(y)); }
+		ful_target("bmi") ful_inline long long andn(long long x, long long y) { return static_cast<long long>(andn(static_cast<unsigned long long>(x), static_cast<unsigned long long>(y))); }
 
 		// bit scan forward
-		ful_inline unsigned int bsf(long long x) { ful_expect(x != 0); return __bsfq(x); }
+		ful_inline unsigned int bsf(long long x) { ful_expect(x != 0); return static_cast<unsigned int>(__bsfq(x)); }
 
 		// bit scan forward
 		ful_inline unsigned int bsf(unsigned long long x) { return bsf(static_cast<long long>(x)); }
 
 		// bit scan reverse
-		ful_inline unsigned int bsr(long long x) { ful_expect(x != 0); return __bsrq(x); }
+		ful_inline unsigned int bsr(long long x) { ful_expect(x != 0); return static_cast<unsigned int>(__bsrq(x)); }
 
 		// bit scan reverse
 		ful_inline unsigned int bsr(unsigned long long x) { return bsr(static_cast<long long>(x)); }
@@ -108,7 +108,7 @@ namespace ful
 		ful_inline long long bswap(long long x) { return __bswapq(x); }
 
 		// byte swap
-		ful_inline unsigned long long bswap(unsigned long long x) { return bswap(static_cast<long long>(x)); }
+		ful_inline unsigned long long bswap(unsigned long long x) { return static_cast<unsigned long long>(bswap(static_cast<long long>(x))); }
 
 		// move data from string to string
 		ful_inline void repmovf(const unsigned long long * src, unsigned long long cnt, unsigned long long * dst)
@@ -128,13 +128,13 @@ namespace ful
 		ful_inline unsigned long long rotl(unsigned long long x, int shift) { return __rolq(x, shift); }
 
 		// rotate left
-		ful_inline long long rotl(long long x, int shift) { return rotl(static_cast<unsigned long long>(x), shift); }
+		ful_inline long long rotl(long long x, int shift) { return static_cast<long long>(rotl(static_cast<unsigned long long>(x), shift)); }
 
 		// rotate right
 		ful_inline unsigned long long rotr(unsigned long long x, int shift) { return __rorq(x, shift); }
 
 		// rotate right
-		ful_inline long long rotr(long long x, int shift) { return rotr(static_cast<unsigned long long>(x), shift); }
+		ful_inline long long rotr(long long x, int shift) { return static_cast<long long>(rotr(static_cast<unsigned long long>(x), shift)); }
 # endif
 
 # if defined(__LP64__)
@@ -142,7 +142,7 @@ namespace ful
 		ful_target("bmi") ful_inline unsigned long andn(unsigned long x, unsigned long y) { return andn(static_cast<unsigned long long>(x), static_cast<unsigned long long>(y)); }
 
 		// not and
-		ful_target("bmi") ful_inline long andn(long x, long y) { return andn(static_cast<unsigned long>(x), static_cast<unsigned long>(y)); }
+		ful_target("bmi") ful_inline long andn(long x, long y) { return static_cast<long>(andn(static_cast<unsigned long>(x), static_cast<unsigned long>(y))); }
 
 		// bit scan forward
 		ful_inline unsigned int bsf(long x) { return bsf(static_cast<long long>(x)); }
@@ -160,7 +160,7 @@ namespace ful
 		ful_inline long bswap(long x) { return bswap(static_cast<long long>(x)); }
 
 		// byte swap
-		ful_inline unsigned long bswap(unsigned long x) { return bswap(static_cast<long>(x)); }
+		ful_inline unsigned long bswap(unsigned long x) { return static_cast<unsigned long>(bswap(static_cast<long>(x))); }
 
 		// move data from string to string
 		ful_inline void repmovf(const unsigned long * src, unsigned long cnt, unsigned long * dst) { repmovf(reinterpret_cast<const unsigned long long *>(src), static_cast<unsigned long long>(cnt), reinterpret_cast<unsigned long long *>(dst)); }
@@ -446,13 +446,13 @@ namespace ful
 		ful_target("movbe") ful_inline short loadbe(const short * x) { return _loadbe_i16(x); }
 
 		// move data after swapping bytes
-		ful_target("movbe") ful_inline unsigned short loadbe(const unsigned short * x) { return loadbe(reinterpret_cast<const short *>(x)); }
+		ful_target("movbe") ful_inline unsigned short loadbe(const unsigned short * x) { return static_cast<unsigned short>(loadbe(reinterpret_cast<const short *>(x))); }
 
 		// move data after swapping bytes
 		ful_target("movbe") ful_inline int loadbe(const int * x) { return _loadbe_i32(x); }
 
 		// move data after swapping bytes
-		ful_target("movbe") ful_inline unsigned int loadbe(const unsigned int * x) { return loadbe(reinterpret_cast<const int *>(x)); }
+		ful_target("movbe") ful_inline unsigned int loadbe(const unsigned int * x) { return static_cast<unsigned int>(loadbe(reinterpret_cast<const int *>(x))); }
 
 		// move data after swapping bytes
 		ful_target("movbe") ful_inline void storebe(short * x, short y) { _storebe_i16(x, y); }
@@ -472,7 +472,7 @@ namespace ful
 		ful_target("movbe") ful_inline long long loadbe(const long long * x) { return _loadbe_i64(x); }
 
 		// move data after swapping bytes
-		ful_target("movbe") ful_inline unsigned long long loadbe(const unsigned long long * x) { return loadbe(reinterpret_cast<const long long *>(x)); }
+		ful_target("movbe") ful_inline unsigned long long loadbe(const unsigned long long * x) { return static_cast<unsigned long long>(loadbe(reinterpret_cast<const long long *>(x))); }
 
 		// move data after swapping bytes
 		ful_target("movbe") ful_inline void storebe(long long * x, long long y) { _storebe_i64(x, y); }
@@ -490,7 +490,7 @@ namespace ful
 		ful_target("movbe") ful_inline long loadbe(const long * x) { return loadbe(reinterpret_cast<const long long *>(x)); }
 
 		// move data after swapping bytes
-		ful_target("movbe") ful_inline unsigned long loadbe(const unsigned long * x) { return loadbe(reinterpret_cast<const long *>(x)); }
+		ful_target("movbe") ful_inline unsigned long loadbe(const unsigned long * x) { return static_cast<unsigned long>(loadbe(reinterpret_cast<const long *>(x))); }
 
 		// move data after swapping bytes
 		ful_target("movbe") ful_inline void storebe(long * x, long y) { storebe(reinterpret_cast<long long *>(x), static_cast<long long>(y)); }
@@ -518,7 +518,7 @@ namespace ful
 		ful_target("bmi2") ful_inline unsigned int bzhi(unsigned int x, unsigned int index) { return _bzhi_u32(x, index); }
 
 		// zero high bits starting with specified bit position
-		ful_target("bmi2") ful_inline int bzhi(int x, unsigned int index) { return bzhi(static_cast<unsigned int>(x), index); }
+		ful_target("bmi2") ful_inline int bzhi(int x, unsigned int index) { return static_cast<int>(bzhi(static_cast<unsigned int>(x), index)); }
 
 		// count the number of leading zero bits
 		ful_target("lzcnt") ful_inline unsigned int lzcnt(unsigned int x) { return _lzcnt_u32(x); }
@@ -530,16 +530,16 @@ namespace ful
 		ful_target("bmi2") ful_inline unsigned int pdep(unsigned int x, unsigned int mask) { return _pdep_u32(x, mask); }
 
 		// parallel bits deposit
-		ful_target("bmi2") ful_inline int pdep(int x, unsigned int mask) { return pdep(static_cast<unsigned int>(x), mask); }
+		ful_target("bmi2") ful_inline int pdep(int x, unsigned int mask) { return static_cast<int>(pdep(static_cast<unsigned int>(x), mask)); }
 
 		// parallel bits extract
 		ful_target("bmi2") ful_inline unsigned int pext(unsigned int x, unsigned int mask) { return _pext_u32(x, mask); }
 
 		// parallel bits extract
-		ful_target("bmi2") ful_inline int pext(int x, unsigned int mask) { return pext(static_cast<unsigned int>(x), mask); }
+		ful_target("bmi2") ful_inline int pext(int x, unsigned int mask) { return static_cast<int>(pext(static_cast<unsigned int>(x), mask)); }
 
 		// return the count of number of bits set to 1
-		ful_target("popcnt") ful_inline unsigned int popcnt(unsigned int x) { return _mm_popcnt_u32(x); }
+		ful_target("popcnt") ful_inline unsigned int popcnt(unsigned int x) { return static_cast<unsigned int>(_mm_popcnt_u32(x)); }
 
 		// return the count of number of bits set to 1
 		ful_target("popcnt") ful_inline unsigned int popcnt(int x) { return popcnt(static_cast<unsigned int>(x)); }
@@ -593,7 +593,7 @@ namespace ful
 		ful_target("bmi2") ful_inline unsigned long long bzhi(unsigned long long x, unsigned int index) { return _bzhi_u64(x, index); }
 
 		// zero high bits starting with specified bit position
-		ful_target("bmi2") ful_inline long long bzhi(long long x, unsigned int index) { return bzhi(static_cast<unsigned long long>(x), index); }
+		ful_target("bmi2") ful_inline long long bzhi(long long x, unsigned int index) { return static_cast<long long>(bzhi(static_cast<unsigned long long>(x), index)); }
 
 		// count the number of leading zero bits
 		ful_target("lzcnt") ful_inline unsigned long long lzcnt(unsigned long long x) { return _lzcnt_u64(x); }
@@ -605,16 +605,16 @@ namespace ful
 		ful_target("bmi2") ful_inline unsigned long long pdep(unsigned long long x, unsigned long long mask) { return _pdep_u64(x, mask); }
 
 		// parallel bits deposit
-		ful_target("bmi2") ful_inline long long pdep(long long x, unsigned long long mask) { return pdep(static_cast<unsigned long long>(x), mask); }
+		ful_target("bmi2") ful_inline long long pdep(long long x, unsigned long long mask) { return static_cast<long long>(pdep(static_cast<unsigned long long>(x), mask)); }
 
 		// parallel bits extract
 		ful_target("bmi2") ful_inline unsigned long long pext(unsigned long long x, unsigned long long mask) { return _pext_u64(x, mask); }
 
 		// parallel bits extract
-		ful_target("bmi2") ful_inline long long pext(long long x, unsigned long long mask) { return pext(static_cast<unsigned long long>(x), mask); }
+		ful_target("bmi2") ful_inline long long pext(long long x, unsigned long long mask) { return static_cast<long long>(pext(static_cast<unsigned long long>(x), mask)); }
 
 		// return the count of number of bits set to 1
-		ful_target("popcnt") ful_inline unsigned long long popcnt(unsigned long long x) { return _mm_popcnt_u64(x); }
+		ful_target("popcnt") ful_inline unsigned long long popcnt(unsigned long long x) { return static_cast<unsigned long long>(_mm_popcnt_u64(x)); }
 
 		// return the count of number of bits set to 1
 		ful_target("popcnt") ful_inline unsigned long long popcnt(long long x) { return popcnt(static_cast<unsigned long long>(x)); }
@@ -631,7 +631,7 @@ namespace ful
 		ful_target("bmi2") ful_inline unsigned long bzhi(unsigned long x, unsigned int index) { return bzhi(static_cast<unsigned long long>(x), index); }
 
 		// zero high bits starting with specified bit position
-		ful_target("bmi2") ful_inline long bzhi(long x, unsigned int index) { return bzhi(static_cast<unsigned long>(x), index); }
+		ful_target("bmi2") ful_inline long bzhi(long x, unsigned int index) { return static_cast<long>(bzhi(static_cast<unsigned long>(x), index)); }
 
 		// count the number of leading zero bits
 		ful_target("lzcnt") ful_inline unsigned long lzcnt(unsigned long x) { return lzcnt(static_cast<unsigned long long>(x)); }
@@ -643,13 +643,13 @@ namespace ful
 		ful_target("bmi2") ful_inline unsigned long pdep(unsigned long x, unsigned long mask) { return pdep(static_cast<unsigned long long>(x), static_cast<unsigned long long>(mask)); }
 
 		// parallel bits deposit
-		ful_target("bmi2") ful_inline long pdep(long x, unsigned long mask) { return pdep(static_cast<unsigned long>(x), mask); }
+		ful_target("bmi2") ful_inline long pdep(long x, unsigned long mask) { return static_cast<long>(pdep(static_cast<unsigned long>(x), mask)); }
 
 		// parallel bits extract
 		ful_target("bmi2") ful_inline unsigned long pext(unsigned long x, unsigned long mask) { return pext(static_cast<unsigned long long>(x), static_cast<unsigned long long>(mask)); }
 
 		// parallel bits extract
-		ful_target("bmi2") ful_inline long pext(long x, unsigned long mask) { return pext(static_cast<unsigned long>(x), mask); }
+		ful_target("bmi2") ful_inline long pext(long x, unsigned long mask) { return static_cast<long>(pext(static_cast<unsigned long>(x), mask)); }
 
 		// return the count of number of bits set to 1
 		ful_target("popcnt") ful_inline unsigned long popcnt(unsigned long x) { return popcnt(static_cast<unsigned long long>(x)); }
